backend/data_service/requirements.txt
google-api-python-client
python-dotenv
mysql-connector-python
validators
requests
flask